move old samplingParams to new dir
